Prevent a hanging server on WSL, when the port is already taken.
The ReqServer uses the REUSEADDR flag in order to avoid having to wait
for the system to clean up old instances if they have crashed without
correctly cleaning themselves up. It allows the port to be re-used by
the new process if the existing bind is no longer actively listening.

*HOWEVER*, the case where the port is genuinely already in use,  is not
implemented correctly on Windows, and rather than preventing the bind,
the process instead 'queues', and waits for the proces to finish. This
leaves the NoDoubleBind test hanging.

To work around this issue, ReqServer will no longer attempt to reuse the
port of a crashed server on WSL